+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Inspections:
 Query Log Viewer:
 
 1. Fix displaying the log viewer when opening a new project into the IDE window.
+1. Don't clear the log viewer when an error is from a previously selected query processor.
 
 ## 2020